qa_crowbarsetup: support to add additional updaterepos after deployment
During regular deployment the function addupdaterepo will add all the repos
from the variable UPDATEREPOS to the PTF repo.
In order to add more repos to the PTF repo after the deployment the function
addupdaterepos can be called again with a parameter (integer > 0).
The first n repos from the variable UPDATEREPOS_EXTRA are added to the PTF repo.
The contents of the repos is always added and never removed.
The value of UPDATEREPOS_EXTRA is not changed.

Therefore:
Calling addupdaterepo twice with the same parameter does not add content to the repo.
Calling it repeatedly with increasing numbers will successively add content.